cronjob/requirements.txt
dnspython==2.7.0
pymongo==4.11.3
python-dotenv==1.1.0
